+++ meanline_axial/__init__.py
@@ -6,6 +6,7 @@ from .math import *
 from .utilities import *
 from .meanline import *
 from .plot_functions import *
+from .configuration import *
 
 
 from .meanline.performance_analysis import *